rename dirs for clarity
