+++ hooks/index.ts
@@ -15,12 +15,12 @@ export { useInterval } from "./helpers/useInterval";
 export { useIsomorphicLayoutEffect } from "./helpers/useIsomorphicLayoutEffect";
 export { useMounted } from "./helpers/useMounted";
 export { useOnChange } from "./helpers/useOnChange";
-export { useApprove } from "./mutations/balances/useApprove";
-export { useExecuteUnstake } from "./mutations/balances/useExecuteUnstake";
-export { useRequestUnstake } from "./mutations/balances/useRequestUnstake";
-export { useStake } from "./mutations/balances/useStake";
-export { useWithdrawAndRestake } from "./mutations/balances/useWithdrawAndRestake";
-export { useWithdrawRewards } from "./mutations/balances/useWithdrawRewards";
+export { useApprove } from "./mutations/staking/useApprove";
+export { useExecuteUnstake } from "./mutations/staking/useExecuteUnstake";
+export { useRequestUnstake } from "./mutations/staking/useRequestUnstake";
+export { useStake } from "./mutations/staking/useStake";
+export { useWithdrawAndRestake } from "./mutations/staking/useWithdrawAndRestake";
+export { useWithdrawRewards } from "./mutations/staking/useWithdrawRewards";
 export { useAcceptReceivedRequestToBeDelegate } from "./mutations/delegation/useAcceptReceivedRequestToBeDelegate";
 export { useCancelSentRequestToBeDelegate } from "./mutations/delegation/useCancelSentRequestToBeDelegate";
 export { useIgnoreReceivedRequestToBeDelegate } from "./mutations/delegation/useIgnoreReceivedRequestToBeDelegate";
@@ -29,10 +29,10 @@ export { useTerminateRelationshipWithDelegate } from "./mutations/delegation/use
 export { useTerminateRelationshipWithDelegator } from "./mutations/delegation/useTerminateRelationshipWithDelegator";
 export { useCommitVotes } from "./mutations/votes/useCommitVotes";
 export { useRevealVotes } from "./mutations/votes/useRevealVotes";
-export { useOutstandingRewards } from "./queries/balances/useOutstandingRewards";
-export { useStakerDetails } from "./queries/balances/useStakerDetails";
-export { useTokenAllowance } from "./queries/balances/useTokenAllowance";
-export { useUnstakedBalance } from "./queries/balances/useUnstakedBalance";
+export { useOutstandingRewards } from "./queries/staking/useOutstandingRewards";
+export { useStakerDetails } from "./queries/staking/useStakerDetails";
+export { useTokenAllowance } from "./queries/staking/useTokenAllowance";
+export { useUnstakedBalance } from "./queries/staking/useUnstakedBalance";
 export { useCommittedVotesForDelegator } from "./queries/delegation/useCommittedVotesForDelegator";
 export { useDelegateToStaker } from "./queries/delegation/useDelegateToStaker";
 export { useDelegatorSetEventsForDelegate } from "./queries/delegation/useDelegatorSetEventsForDelegate";
@@ -51,6 +51,6 @@ export { useEncryptedVotes } from "./queries/votes/useEncryptedVotes";
 export { useHasActiveVotes } from "./queries/votes/useHasActiveVotes";
 export { usePastVotes } from "./queries/votes/usePastVotes";
 export { useRevealedVotes } from "./queries/votes/useRevealedVotes";
-export { useUnstakeCoolDown } from "./queries/balances/useUnstakeCoolDown";
+export { useUnstakeCoolDown } from "./queries/staking/useUnstakeCoolDown";
 export { useUpcomingVotes } from "./queries/votes/useUpcomingVotes";
 export { useVoteTransactionHashes } from "./queries/votes/useVoteTransactionHashes";
