+++ hooks/queries/staking/useStakerDetails.ts
