+++ hooks/queries/staking/useTokenAllowance.ts
